refactor(DeleteTag): rename the folder from 'RemoveTag' to 'DeleteTag' for increase cohesion
