+++ src/ToDoManager.Application/Tags/Commands/DeleteTag/DeleteTagCommand.cs
@@ -2,6 +2,6 @@
 using MediatR;
 using ToDoManager.Domain.Tags.ValueObjects;
 
-namespace ToDoManager.Application.Tags.Commands.RemoveTag;
+namespace ToDoManager.Application.Tags.Commands.DeleteTag;
 
 public record DeleteTagCommand(TagId TagId) : IRequest<ErrorOr<Unit>>;
